refactor user_email_input
as of the changes from #24, I feel it is a better approach to push the editors from the user picker individually to enteredEditors in collection-form-screen.
This also solves the problem, that it was either impossible to remove someone or all editors were automatically removed.

also I find user_picker is a better name than user_email_input