show number of current item
this adds a label to the text input of the group-body-form.

where if two items are already in the group body the label would show: "#3".

Although this is nice, I am not yet sure if it's not overkill or will be disadvantageous when connection is bad, so it might only be a temporary feature